[WiP] Added GameLoop
Game loop with current design looks like godclass.
Should be broken down to few smaller classes.
Builder pattern might be a good idea for a constructor,
given the amount of parameters.
