+++ src/main/java/SnakeFX_Game.java
@@ -1,11 +1,9 @@
 package main.java;
 
 import javafx.application.Application;
-import javafx.event.EventHandler;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.input.KeyEvent;
 import javafx.stage.Stage;
 
 /**
@@ -16,8 +14,8 @@ public class SnakeFX_Game extends Application {
     
     @Override
     public void start(Stage stage) throws Exception {
-
-        Parent root = FXMLLoader.load(getClass().getResource("/fxml/FXMLDocument.fxml"));
+        
+        Parent root = FXMLLoader.load(getClass().getResource("/fxml/FXMLDocument.fxml"));        
         
         Scene scene = new Scene(root);
