+++ src/main/java/food/FoodPOJO.java
@@ -1,0 +1,10 @@
+package main.java.food;
+
+public class FoodPOJO {
+    private double x;
+    private double y;
+
+    public FoodPOJO() {
+
+    }
+}
