+++ src/main/java/view/food/FoodRectangle.java
@@ -1,0 +1,8 @@
+package main.java.view.food;
+
+public class FoodRectangle {
+
+    public FoodRectangle() {
+
+    }
+}
